requirements.txt
flask
transformers
torch
torchvision
pillow
opencv-python-headless